Refactor the webdataset loader code, move classes to individual files, resolve dependencies
