+++ src/megatron/energon/flavors/ocr.py
@@ -7,7 +7,7 @@ from typing import List, Optional, Union
 import torch
 
 from megatron.energon.flavors.base_dataset import Sample
-from megatron.energon.flavors.base_webdataset import DefaultDecoderWebdataset
+from megatron.energon.flavors.webdataset import DefaultDecoderWebdataset
 
 
 @dataclass
