+++ src/megatron/energon/flavors/text.py
@@ -4,7 +4,7 @@
 from dataclasses import dataclass
 
 from megatron.energon.flavors.base_dataset import Sample
-from megatron.energon.flavors.base_webdataset import DefaultDecoderWebdataset
+from megatron.energon.flavors.webdataset import DefaultDecoderWebdataset
 
 
 @dataclass
