+++ src/megatron/energon/flavors/vid_qa.py
@@ -7,7 +7,7 @@ from typing import List, Optional
 import torch
 
 from megatron.energon.flavors.base_dataset import Sample
-from megatron.energon.flavors.base_webdataset import DefaultDecoderWebdataset, VideoData
+from megatron.energon.flavors.webdataset import DefaultDecoderWebdataset, VideoData
 
 
 @dataclass
