+++ src/megatron/energon/flavors/webdataset/__init__.py
@@ -1,0 +1,21 @@
+from megatron.energon.flavors.webdataset.base_webdataset import BaseWebdataset
+from megatron.energon.flavors.webdataset.default_decoder_webdataset import (
+    DefaultDecoderWebdataset,
+    ImageDecoder,
+)
+from megatron.energon.flavors.webdataset.default_generic_webdataset import DefaultGenericWebdataset
+from megatron.energon.flavors.webdataset.empty_dataset_error import EmptyDatasetError
+from megatron.energon.flavors.webdataset.merged_webdataset import MergedWebdataset
+from megatron.energon.flavors.webdataset.metadata import WebdatasetMeta
+from megatron.energon.flavors.webdataset.video_data import VideoData
+
+__all__ = [
+    "BaseWebdataset",
+    "EmptyDatasetError",
+    "VideoData",
+    "DefaultDecoderWebdataset",
+    "DefaultGenericWebdataset",
+    "ImageDecoder",
+    "WebdatasetMeta",
+    "MergedWebdataset",
+]
