+++ src/megatron/energon/flavors/webdataset/empty_dataset_error.py
@@ -1,0 +1,2 @@
+class EmptyDatasetError(Exception):
+    """Raised when a dataset is empty."""
