+++ src/megatron/energon/flavors/webdataset/structs.py
@@ -45,9 +45,9 @@ class FilteredSample(TypedDict):
     #: The base name of the shard file e.g. "shard_000"
     __shard__: str
     #: Globally unique key to restore a sample from disk.
-    #: For example `("shard_000", 123)` would restore
+    #: For example `("Webdataset", "shard_000", 123)` would restore
     #: the sample from "shard_000.tar" at byte offset 123.
-    __restore_key__: Tuple[str, int]
+    __restore_key__: Tuple[str, str, int]
 
 
 def reraise_exception(exc: Exception, key: Optional[str]) -> None:
@@ -55,3 +55,18 @@ def reraise_exception(exc: Exception, key: Optional[str]) -> None:
         raise Exception(f"For sample {key!r}") from exc
     else:
         raise
+
+
+@dataclass
+class ShardState:
+    """
+    Represents the iteration state of a single shard. I.e. contains a reference to the shard and the
+    reader's offset.
+    """
+
+    #: The shard of this state
+    shard: ShardInfo
+    #: Sample offset relative to the sub-shard file offset
+    offset: int
+    #: Byte offset relative to the sub-shard file byte offset
+    byte_offset: int
