+++ src/megatron/energon/tools/info.py
@@ -1,19 +1,15 @@
 # Copyright (c) 2024, NVIDIA CORPORATION.
 # SPDX-License-Identifier: BSD-3-Clause
 
+from importlib import import_module
 from io import StringIO
 from typing import Type
-import click
-from importlib import import_module
-
-from megatron.energon.dataset_config import MAIN_FOLDER_NAME
-from megatron.energon.epathlib import EPath
-
-
 
 import click
 import yaml
 
+from megatron.energon.dataset_config import MAIN_FOLDER_NAME
+from megatron.energon.epathlib import EPath
 from megatron.energon.flavors.base_dataset import BaseCoreDataset, Sample
 
 fmt = """
@@ -94,10 +90,7 @@ def command(
             path=str(path),
             samples_count=samples_count,
             samples_size=fmt_size(
-                sum(
-                    (path / split_name).size()
-                    for split_name in info_config["shard_counts"].keys()
-                )
+                sum((path / split_name).size() for split_name in info_config["shard_counts"].keys())
             ),
             shards_count=len(info_config["shard_counts"]),
             class_name=class_name,
